Moved some validators out of validationHelpers
