+++ app/models/creaturetype.js
@@ -1,15 +1,8 @@
 const { Model } = require('sequelize');
 
 const {
-  isArrayOfActionPatterns,
-  isArrayOfAlignmentStrings,
   isArrayOfStrings,
   isArrayOfLabeledDescriptions,
-  isArrayOfSkillObjects,
-  isValidResistancesObject,
-  isValidSavingThrowsObject,
-  isValidSpellSlotArray,
-  isValidInnateSpellArray,
   isValidStat,
 } = require('../services/validationHelpers');
 
@@ -53,7 +46,17 @@ module.exports = (sequelize, DataTypes) => {
     alignment: {
       type: DataTypes.JSON,
       validate: {
-        isArrayOfAlignmentStrings,
+        isArrayOfAlignmentStrings(array) {
+          if (array === null) return;
+          isArrayOfStrings(array);
+          if (typeof array === 'string') array = JSON.parse(array);
+          if (array.length !== 2) throw new Error('alignment array must be length two');
+          if (!(array[0] === 'chaotic' || array[0] === 'neutral' || array[0] === 'lawful')
+            || !(array[1] === 'evil' || array[1] === 'neutral' || array[1] === 'good')
+          ) {
+            throw new Error('first alignment must be on ethical axis, second alignment on moral axis');
+          }
+        },
       },
     },
     armorId: DataTypes.INTEGER,
@@ -163,19 +166,71 @@ module.exports = (sequelize, DataTypes) => {
     savingThrows: {
       type: DataTypes.JSON,
       validate: {
-        isValidSavingThrowsObject,
+        isSavingThrowsObject(object) {
+          if (object === null) return;
+          if (typeof object === 'string') {
+            object = JSON.parse(object);
+          }
+          const objectKeys = Object.keys(object);
+          if (objectKeys.length !== 6
+            || !objectKeys.includes('str')
+            || !objectKeys.includes('dex')
+            || !objectKeys.includes('str')
+            || !objectKeys.includes('int')
+            || !objectKeys.includes('wis')
+            || !objectKeys.includes('cha')
+            || object.str < 0 || object.str > 30
+            || object.dex < 0 || object.dex > 30
+            || object.con < 0 || object.con > 30
+            || object.int < 0 || object.int > 30
+            || object.wis < 0 || object.wis > 30
+            || object.cha < 0 || object.cha > 30
+          ) throw new Error('saving throws object must contain each ability as a key');
+        },
       },
     },
     skills: {
       type: DataTypes.JSON,
       validate: {
-        isArrayOfSkillObjects,
+        isArrayOfSkillObjects(array) {
+          if (array === null) return;
+          if (typeof array === 'string') {
+            array = JSON.parse(array);
+          }
+          if (!array.isArray()) throw new Error('skill array must be an array');
+          if (array.length <= 0) throw new Error('skill array should not be length 0');
+          array.forEach((object) => {
+            if (typeof object !== 'object') throw new Error('skill array must contain objects');
+            const objectKeys = Object.keys(object);
+            if (
+              objectKeys.length !== 2
+              || !objectKeys.includes('skill')
+              || !objectKeys.includes('value')
+              || typeof object.skill !== 'string'
+              || typeof object.value !== 'number'
+            ) throw new Error('skill array object must contain a skill, value pair');
+          });
+        },
       },
     },
     resistances: {
       type: DataTypes.JSON,
       validate: {
-        isValidResistancesObject,
+        isResistancesObject(object) {
+          if (object === null) return;
+          if (typeof object === 'string') {
+            object = JSON.parse(object);
+          }
+          const objectKeys = Object.keys(object);
+          if (objectKeys.length !== 6
+            || !objectKeys.includes('resistant')
+            || !objectKeys.includes('vulnerable')
+            || !objectKeys.includes('immune')
+          ) throw new Error('resistances object must have resistant, vulnerable, and immune as keys');
+          isArrayOfStrings(object.resistant);
+          isArrayOfStrings(object.vulnerable);
+          isArrayOfStrings(object.immune);
+        },
       },
     },
     senses: {
@@ -234,19 +289,120 @@ module.exports = (sequelize, DataTypes) => {
     spellSlots: {
       type: DataTypes.JSON,
       validate: {
-        isValidSpellSlotArray,
+        /**
+         * for example [0,5,2,0,0,0,0,0,0,0]
+         * @param {Array} array
+         * @throws
+         */
+        isSpellSlotArray(array) {
+          if (array === null) return;
+          if (typeof array === 'string') {
+            array = JSON.parse(array);
+          }
+          if (!array.isArray()) throw new Error('spell slot array must be an array');
+          if (array.length !== 10) throw new Error('spell slot array should be length 10');
+          if (array[0] !== 0) throw new Error('index 0 of spell slot array should be 0');
+          array.forEach((number) => {
+            if (typeof number !== 'number' || number < 0 || number > 10) {
+              throw new Error('spell slot array must contain only numbers between 0 and 10');
+            }
+          });
+        },
       },
     },
     innateSpells: {
       type: DataTypes.JSON,
       validate: {
-        isValidInnateSpellArray,
+        /**
+         * innateSpell: {
+         *   spellId: integer,
+         *   perDay: [0-9], // perDay 0 => "at will"
+         *   restrictions: string,
+         * }
+         * @param {Array} array
+         * @throws
+         */
+        isInnateSpellArray(array) {
+          if (array === null) return;
+          if (typeof array === 'string') {
+            array = JSON.parse(array);
+          }
+          if (!array.isArray()) throw new Error('innate spell array must be an array');
+          array.forEach((innateSpell) => {
+            if (typeof innateSpell !== 'object') throw new Error('innate spell array must be an array of objects');
+            const objectKeys = Object.keys(innateSpell);
+            if (
+              objectKeys.length !== 3
+              || !objectKeys.includes('spellId')
+              || !objectKeys.includes('perDay')
+              || !objectKeys.includes('restrictions')
+              || typeof innateSpell.spellId !== 'number'
+              || innateSpell.spellId < 1
+              || typeof innateSpell.perDay !== 'number'
+              || innateSpell.perDay < 0
+              || typeof innateSpell.restrictions !== 'string'
+            ) throw new Error('innate spell object must contain a spellId, a perDay, and a restrictions');
+          });
+        },
       },
     },
     actionPatterns: {
       type: DataTypes.JSON,
       validate: {
-        isArrayOfActionPatterns,
+        /**
+         * actionPatterns: [
+         *   "actionPattern" [ "action" {
+         *     other: string,        // If not '', an action other than a weapon or spell attack
+         *     restrictions: string, // Any restrictions the action has
+         *     spellId: [0-inf],     // id of a spell
+         *     times: [1-9],         // number of times to attack with the weapon
+         *     weaponId: [0-inf],    // id of a weapon
+         *                           // Disallow non-zero spellId and non-zero weaponId
+         *                           // If spellId is non-zero, times should be 1.
+         *   }],
+         * ]
+         */
+        isArrayOfActionPatterns(array) {
+          if (typeof array === 'string') {
+            array = JSON.parse(array);
+          }
+          if (!array.isArray()) throw new Error('action patterns array must be an array');
+          if (array.length <= 0) throw new Error('action patterns array should not be length 0');
+          array.forEach((actionPattern) => {
+            if (typeof actionPattern !== 'object' || !actionPattern.isArray()) throw new Error('action patterns array must contain arrays');
+            actionPattern.forEach((object) => {
+              if (typeof object !== 'object') throw new Error('action pattern must be an array of action objects');
+              const objectKeys = Object.keys(object);
+              if (
+                objectKeys.length !== 5
+                || !objectKeys.includes('other')
+                || !objectKeys.includes('restrictions')
+                || !objectKeys.includes('spellId')
+                || !objectKeys.includes('times')
+                || !objectKeys.includes('weaponId')
+                || typeof object.other !== 'string'
+                || typeof object.restrictions !== 'string'
+                || typeof object.spellId !== 'number'
+                || typeof object.times !== 'number'
+                || typeof object.weaponId !== 'number'
+              ) {
+                throw new Error('action object must have keys other, restrictions, spellId, times, and weaponId');
+              }
+              if (object.other.length === 0 && object.spellId === 0 && object.weaponId === 0) {
+                throw new Error('action must be one of other, spell, or weapon attack');
+              }
+              if (object.spellId !== 0 && object.weaponId !== 0) {
+                throw new Error('action cannot be both a spell and a weapon attack');
+              }
+              if (object.other.length > 0 && (object.spellId !== 0 || object.weaponId !== 0)) {
+                throw new Error('action cannot be both other and a spell or weapon attack');
+              }
+              if (object.spellId !== 0 && object.times !== 1) {
+                throw new Error('spell actions cannot be used multiple times');
+              }
+            });
+          });
+        },
       },
     },
     legendaryActions: {
@@ -276,6 +432,11 @@ module.exports = (sequelize, DataTypes) => {
   }, {
     sequelize,
     modelName: 'CreatureType',
+    validate: {
+      maxHPBoundedByHitDice() {
+        if (this.maxHP > (this.numDice * (this.hitDie + this.con))) throw new Error('maxHP field cannot be greater than hit dice allow');
+      },
+    },
   });
   return CreatureType;
 };
